Fixed pie tooltip
